Reorganize app folder structure
