+++ lib/widgets/molecules/detail.dart
